quiz-api/requirements.txt
autopep8==2.3.2
blinker==1.9.0
click==8.1.3
colorama==0.4.5
Flask==2.3.3
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
importlib-metadata==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.14.0
PyJWT==2.5.0
six==1.17.0
SQLAlchemy==2.0.43
toml==0.10.2
typing_extensions==4.15.0
Werkzeug==2.3.8
zipp==3.10.0
